might need to convert pdf for easier view

+++ downloadimg.py
@@ -2,6 +2,7 @@ import os
 import requests
 from tqdm import tqdm
 
+
 """
 pre made download function from the internet to download files and got inbuilt function
 """
@@ -19,13 +20,20 @@ def download(url, pathname):
     file_size = int(response.headers.get("Content-Length", 0))
     # get the file name
     filename = os.path.join(pathname, url.split("/")[-1])
-    # progress bar, changing the unit to bytes instead of iteration (default by tqdm)
-    progress = tqdm(response.iter_content(1024), f"Downloading {filename}", total=file_size, unit="B", unit_scale=True, unit_divisor=1024)
-    with open(filename, "wb") as f:
-        for data in progress:
-            # write data read to the file
-            f.write(data)
-            # update the progress bar manually
-            progress.update(len(data))
+    if not os.path.exists(filename):
+        # progress bar, changing the unit to bytes instead of iteration (default by tqdm)
+        progress = tqdm(response.iter_content(1024), f"Downloading {filename}", total=file_size, unit="B", unit_scale=True, unit_divisor=1024)
+        with open(filename, "wb") as f:
+            for data in progress:
+                # write data read to the file
+                f.write(data)
+                # update the progress bar manually
+                progress.update(len(data))
+    else:
+        print(filename+" exists")
+
+
+
+
